+++ apps/portfolio/components/social-links/SocialLinks.vue
@@ -1,6 +1,18 @@
 <script lang="ts" setup>
 import { PropType } from 'vue'
-import { SocialLink } from './consts'
+
+export interface StoryLink {
+  id: string
+  url: string
+  linktype: string
+  fieldtype: string
+  cached_url: string
+}
+
+export interface SocialLink {
+  icon: string
+  url: StoryLink
+}
 
 const props = defineProps({
   size: {
