--- apps/portfolio/components/social-links/consts.ts
@@ -1,12 +1,0 @@
-export interface StoryLink {
-  id: string
-  url: string
-  linktype: string
-  fieldtype: string
-  cached_url: string
-}
-
-export interface SocialLink {
-  icon: string
-  url: StoryLink
-}
