+++ apps/portfolio/package.json
@@ -31,9 +31,7 @@
     "lite-youtube-embed": "^0.2.0",
     "nuxt": "3.0.0-rc.3",
     "swiper": "^8.1.4",
-    "vite-svg-loader": "^3.3.0"
-  },
-  "dependencies": {
+    "vite-svg-loader": "^3.3.0",
     "velocity-animate": "^1.5.2"
   }
 }
